+++ docs/modules.rst
@@ -37,6 +37,16 @@ Ellipse Fit
 .. automodule:: orbitdeterminator.kep_determination.ellipse_fit
    :members:
 
+Least squares
+~~~~~~~~~~~~~~~~~~~~
+.. automodule:: orbitdeterminator.kep_determination.least_squares
+   :members:
+
+Gauss method
+~~~~~~~~~~~~~~~~~~~~
+.. automodule:: orbitdeterminator.kep_determination.gauss_method
+   :members:
+
 Propagation:
 ------------
