+++ orbitdeterminator/kep_determination/least_squares.py
@@ -20,109 +20,8 @@ import numpy as np
 import matplotlib.pyplot as plt
 from scipy.optimize import minimize
 from scipy.optimize import least_squares
-from ellipse_fit import determine_kep, __read_file
-
-# convention:
-# a: semi-major axis
-# e: eccentricity
-# eps: mean longitude at epoch
-# taup: time of pericenter passage
-# Euler angles:
-# I: inclination
-# Omega: longitude of ascending node
-# omega: argument of pericenter
-
-#rotation about the z-axis about an angle `ang`
-def rotz(ang):
-    cos_ang = np.cos(ang)
-    sin_ang = np.sin(ang)
-    return np.array(((cos_ang,-sin_ang,0.0), (sin_ang, cos_ang,0.0), (0.0,0.0,1.0)))
-
-#rotation about the x-axis about an angle `ang`
-def rotx(ang):
-    cos_ang = np.cos(ang)
-    sin_ang = np.sin(ang)
-    return np.array(((1.0,0.0,0.0), (0.0,cos_ang,-sin_ang), (0.0,sin_ang,cos_ang)))
-
-#rotation from the orbital plane to the inertial frame
-#it is composed of the following rotations, in that order:
-#1) rotation about the z axis about an angle `omega` (argument of pericenter)
-#2) rotation about the x axis about an angle `I` (inclination)
-#3) rotation about the z axis about an angle `Omega` (longitude of ascending node)
-def orbplane2frame_(omega,I,Omega):
-    P2_mul_P3 = np.matmul(rotx(I),rotz(omega))
-    return np.matmul(rotz(Omega),P2_mul_P3)
-
-def orbplane2frame(x):
-    return orbplane2frame_(x[0],x[1],x[2])
-
-# get Keplerian range
-def kep_r_(a, e, f):
-    return a*(1.0-e**2)/(1.0+e*np.cos(f))
-
-def kep_r(x):
-    return kep_r_(x[0],x[1],x[2])
-
-# get cartesian positions wrt orbital plane
-def xyz_orbplane_(a, e, f):
-    r = kep_r_(a, e, f)
-    return np.array((r*np.cos(f),r*np.sin(f),0.0))
-
-def xyz_orbplane(x):
-    return xyz_orbplane_(x[0],x[1],x[2])
-
-# get cartesian positions wrt inertial frame from orbital elements
-def xyz_frame_(a,e,f,omega,I,Omega):
-    return np.matmul( orbplane2frame_(omega,I,Omega) , xyz_orbplane_(a, e, f) )
-
-def xyz_frame(x):
-    return np.matmul( orbplane2frame(x[3:6]) , xyz_orbplane(x[0:3]) )
-
-# get mean motion from mass parameter (mu) and semimajor axis (a)
-def meanmotion(mu,a):
-    return np.sqrt(mu/(a**3))
-
-# get mean anomaly from mean motion (n), time (t) and time of pericenter passage (taup)
-def meananomaly(n, t, taup):
-    return np.mod(n*(t-taup), 2.0*np.pi)
-
-# compute eccentric anomaly (E) from eccentricity (e) and mean anomaly (M)
-def eccentricanomaly(e,M):
-    M0 = np.mod(M,2*np.pi)
-    E0 = M0 + np.sign(np.sin(M0))*0.85*e #Murray-Dermotts' initial estimate
-    # successive approximations via Newtons' method
-    for i in range(0,4):
-        #TODO: implement modified Newton's method for Kepler's equation (Murray-Dermott)
-        Eans = E0 - (E0-e*np.sin(E0)-M0)/(1.0-e*np.cos(E0))
-        E0 = Eans
-    return E0
-
-# compute true anomaly (f) from eccentricity (e) and eccentric anomaly (E)
-def trueanomaly(e,E):
-    Enew = np.mod(E,2.0*np.pi)
-    return 2.0*np.arctan(  np.sqrt((1.0+e)/(1.0-e))*np.tan(Enew/2)  )
-
-# compute eccentric anomaly (E) from eccentricity (e) and true anomaly (f)
-def truean2eccan(e, f):
-    fnew = np.mod(f,2.0*np.pi)
-    return 2.0*np.arctan(  np.sqrt((1.0-e)/(1.0+e))*np.tan(fnew/2)  )
-
-# compute true anomaly from eccentricity and mean anomaly
-def meanan2truean(e,M):
-    return trueanomaly(e, eccentricanomaly(e, M))
-
-# compute true anomaly from time, a, e, mu and taup
-def time2truean(a, e, mu, t, taup):
-    return meanan2truean(e, meananomaly(meanmotion(mu, a), t, taup))
-
-# compute cartesian positions (x,y,z) at time t
-# for mass parameter mu from orbital elements a, e, taup, I, omega, Omega
-def orbel2xyz(t, mu, a, e, taup, omega, I, Omega):
-
-    # compute true anomaly at time t
-    f = time2truean(a, e, mu, t, taup)
-    # get cartesian positions wrt inertial frame from orbital elements
-    return xyz_frame_(a, e, f, omega, I, Omega)
+from orbitdeterminator.kep_determination.ellipse_fit import determine_kep, __read_file
+from orbitdeterminator.kep_determination.gauss_method import *
 
 # compute residuals vector, with Earth's grav parameter as to-be-fitted variable
 def res_vec(x, my_data):
@@ -154,33 +53,260 @@ def Q(x, my_data):
         Q0 = Q0 + np.linalg.norm(xyz_res, ord=2)/my_data.shape[0]
     return Q0
 
-#########################
-
-# Earth's mass parameter in appropriate units:
-mu_Earth = 398600.435436E9 # m^3/seg^2
-#Earth's radius in appropriate units:
-# R_Earth =  6378136.3 #m
-#minimal acceptable altitude for satellites (150 km)??
-#maximal acceptable altitude for satellites (150 km)??
-
-#write file name of data:
-fname = '../orbit.csv'
-
-# load observational data:
-data = np.loadtxt(fname,skiprows=1,usecols=(0,1,2,3))
-
 # cost function of only one argument, x
 # due to optimization of processing time, only the first 2,000 data points are used
 # nevertheless, this is enough to improve the solution
 def QQ(x):
     return Q(x, data)
 
-# generate vector of initial guess of orbital elements:
-# values written below correspond to solution of ellipse_fit.py for the same file
+def gauss_LS_sat(body_fname_str, body_name_str, obs_arr, r2_root_ind_vec=None, obs_arr_ls=None, gaussiters=0, plot=True):
+    """Earth satellites orbit determination high-level function from
+    IOD-formatted ra/dec tracking data. IOD angle subformat 2 is assumed.
+    Preliminary orbit determination via Gauss method is performed.
+    Roots of 8-th order Gauss polynomial are computed using np.roots function.
+    Note that if `r2_root_ind_vec` is not specified by the user, then the first
+    positive root returned by np.roots is used by default.
+
+       Args:
+           body_fname_str (string): path to IOD-formatted observation data file
+           body_name_str (string): user-defined name of satellite
+           obs_arr (int vector): line numbers in data file to be processed in Gauss preliminary orbit determination
+           r2_root_ind_vec (1xlen(obs_arr) int array): indices of Gauss polynomial roots.
+           obs_arr (int vector): line numbers in data file to be processed in least-squares fit
+           gaussiters (int): number of refinement iterations to be performed
+           plot (bool): if True, plots data.
+
+       Returns:
+           x (tuple): set of Keplerian orbital elements (a, e, taup, omega, I, omega, T)
+    """
+    # load IOD data for a given satellite
+    iod_object_data = load_iod_data(body_fname_str)
+
+    #load data of listed observatories (longitude, latitude, elevation)
+    sat_observatories_data = load_sat_observatories_data('sat_tracking_observatories.txt')
+
+    #get preliminary orbit using Gauss method
+    #q0 : a, e, taup, I, W, w, T
+    q0 = np.array(gauss_method_sat(body_fname_str, body_name_str, obs_arr, r2_root_ind_vec=r2_root_ind_vec, refiters=gaussiters, plot=False))
+    x0 = q0[0:6]
+    x0[3:6] = np.deg2rad(x0[3:6])
+
+    # if obs_arr_ls was not specified, then read whole data set:
+    if obs_arr_ls is None:
+        obs_arr_ls = np.array(range(1, len(iod_object_data)+1))
+
+    rov = radec_obs_vec_sat(obs_arr_ls, iod_object_data)
+    rv0 = radec_res_vec_rov_sat(x0, obs_arr_ls, iod_object_data, sat_observatories_data, rov)
+    Q0 = np.linalg.norm(rv0, ord=2)/len(rv0)
+
+    Q_ls = least_squares(radec_res_vec_rov_sat, x0, args=(obs_arr_ls, iod_object_data, sat_observatories_data, rov), method='lm', xtol=1e-13)
+
+    print('\n*** ORBIT DETERMINATION: LEAST-SQUARES FIT ***')
+
+    print('\nINFO: scipy.optimize.least_squares exited with code', Q_ls.status)
+    print(Q_ls.message,'\n')
+
+    tv_star, rv_star = t_radec_res_vec_sat(Q_ls.x, obs_arr_ls, iod_object_data, sat_observatories_data, rov)
+    Q_star = np.linalg.norm(rv_star, ord=2)/len(rv_star)
+
+    print('Total residual evaluated at averaged Gauss solution: ', Q0)
+    print('Total residual evaluated at least-squares solution: ', Q_star, '\n')
+
+    print('Observational arc:')
+    print('Number of observations: ', len(obs_arr_ls))
+    print('First observation (UTC) : ', Time(tv_star[0], format='jd').iso)
+    print('Last observation (UTC) : ', Time(tv_star[-1], format='jd').iso)
+
+    n_num = meanmotion(mu_Earth, Q_ls.x[0])
+
+    print('\nORBITAL ELEMENTS (EQUATORIAL): a, e, taup, omega, I, Omega, T')
+    print('Semi-major axis (a):                 ', Q_ls.x[0], 'km')
+    print('Eccentricity (e):                    ', Q_ls.x[1])
+    print('Time of pericenter passage (tau):    ', Time(Q_ls.x[2], format='jd').iso, 'JDUTC')
+    print('Argument of pericenter (omega):      ', np.rad2deg(Q_ls.x[3]), 'deg')
+    print('Inclination (I):                     ', np.rad2deg(Q_ls.x[4]), 'deg')
+    print('Longitude of Ascending Node (Omega): ', np.rad2deg(Q_ls.x[5]), 'deg')
+    print('Orbital period (T):                  ', 2.0*np.pi/n_num/60.0, 'min')
+
+    # PLOT
+    if plot:
+        ra_res_vec = np.rad2deg(rv_star[0::2])*(3600.0)
+        dec_res_vec = np.rad2deg(rv_star[1::2])*(3600.0)
+        t_plot = (tv_star-tv_star[0])*86400.0
+
+        f, axarr = plt.subplots(2, sharex=True)
+        axarr[0].set_title('Gauss + LS fit residuals: RA, Dec')
+        axarr[0].scatter(t_plot, ra_res_vec, label='delta RA (\")', marker='+')
+        axarr[0].set_ylabel('RA (\")')
+        axarr[1].scatter(t_plot, dec_res_vec, label='delta Dec (\")', marker='+')
+        axarr[1].set_xlabel('time (UTC seconds since first obs)')
+        axarr[1].set_ylabel('Dec (\")')
+        plt.show()
+
+        npoints = 500
+        theta_vec = np.linspace(0.0, 2.0*np.pi, npoints)
+        x_orb_vec = np.zeros((npoints,))
+        y_orb_vec = np.zeros((npoints,))
+        z_orb_vec = np.zeros((npoints,))
+
+        for i in range(0,npoints):
+            x_orb_vec[i], y_orb_vec[i], z_orb_vec[i] = xyz_frame2(Q_ls.x[0], Q_ls.x[1], theta_vec[i], Q_ls.x[3], Q_ls.x[4], Q_ls.x[5])
+
+        ax = plt.axes(aspect='equal', projection='3d')
+
+        # Earth-centered orbits: satellite orbit and geocenter
+        ax.scatter3D(0.0, 0.0, 0.0, color='blue', label='Earth')
+        ax.plot3D(x_orb_vec, y_orb_vec, z_orb_vec, 'red', linewidth=0.5, label=body_name_str+' orbit')
+        plt.legend()
+        ax.set_xlabel('x (km)')
+        ax.set_ylabel('y (km)')
+        ax.set_zlabel('z (km)')
+        xy_plot_abs_max = np.max((np.amax(np.abs(ax.get_xlim())), np.amax(np.abs(ax.get_ylim()))))
+        ax.set_xlim(-xy_plot_abs_max, xy_plot_abs_max)
+        ax.set_ylim(-xy_plot_abs_max, xy_plot_abs_max)
+        ax.set_zlim(-xy_plot_abs_max, xy_plot_abs_max)
+        ax.legend(loc='center left', bbox_to_anchor=(1.04,0.5))
+        ax.set_title('Satellite orbit (Gauss+LS): '+body_name_str)
+        plt.show()
+
+    return Q_ls.x[0], Q_ls.x[1], Time(Q_ls.x[2], format='jd'), np.rad2deg(Q_ls.x[3]), np.rad2deg(Q_ls.x[4]), np.rad2deg(Q_ls.x[5]), 2.0*np.pi/n_num/60.0
+
+def gauss_LS_mpc(body_fname_str, body_name_str, obs_arr, r2_root_ind_vec=None, obs_arr_ls=None, gaussiters=0, plot=True):
+    """Minor planets orbit determination high-level function from MPC-formatted
+    ra/dec tracking data. Preliminary orbit determination via Gauss method is
+    performed. Roots of 8-th order Gauss polynomial are computed using np.roots
+    function. Note that if `r2_root_ind_vec` is not specified by the user, then
+    the first positive root returned by np.roots is used by default.
+
+       Args:
+           body_fname_str (string): path to MPC-formatted observation data file
+           body_name_str (string): user-defined name of minor planet
+           obs_arr (int vector): line numbers in data file to be processed in Gauss preliminary orbit determination
+           r2_root_ind_vec (1xlen(obs_arr) int array): indices of Gauss polynomial roots.
+           obs_arr (int vector): line numbers in data file to be processed in least-squares fit
+           gaussiters (int): number of refinement iterations to be performed
+           plot (bool): if True, plots data.
+
+       Returns:
+           x (tuple): set of Keplerian orbital elements (a, e, taup, omega, I, omega, T)
+    """
+    # load MPC data for a given NEA
+    mpc_object_data = load_mpc_data(body_fname_str)
+
+    #load MPC data of listed observatories (longitude, parallax constants C, S)
+    mpc_observatories_data = load_mpc_observatories_data('mpc_observatories.txt')
+
+    #x0 : a, e, taup, I, W, w
+    x0 = np.array(gauss_method_mpc(body_fname_str, body_name_str, obs_arr, r2_root_ind_vec=r2_root_ind_vec, refiters=gaussiters, plot=False))
+
+    x0[3:6] = np.deg2rad(x0[3:6])
+
+    # if obs_arr_ls was not specified, then read whole data set:
+    if obs_arr_ls is None:
+        obs_arr_ls = np.array(range(1, len(mpc_object_data)+1))
+
+    rov = radec_obs_vec_mpc(obs_arr_ls, mpc_object_data)
+
+    rv0 = radec_res_vec_rov_mpc(x0, obs_arr_ls, mpc_object_data, mpc_observatories_data, rov)
+    Q0 = np.linalg.norm(rv0, ord=2)/len(rv0)
+
+    print('\n*** ORBIT DETERMINATION: LEAST-SQUARES FIT ***')
+
+    Q_ls = least_squares(radec_res_vec_rov_mpc, x0, args=(obs_arr_ls, mpc_object_data, mpc_observatories_data, rov), method='lm')
+
+    print('\nINFO: scipy.optimize.least_squares exited with code ', Q_ls.status)
+    print(Q_ls.message,'\n')
+
+    tv_star, rv_star = t_radec_res_vec_mpc(Q_ls.x, obs_arr_ls, mpc_object_data, mpc_observatories_data)
+    Q_star = np.linalg.norm(rv_star, ord=2)/len(rv_star)
+
+    print('Total residual evaluated at averaged Gauss solution: ', Q0)
+    print('Total residual evaluated at least-squares solution: ', Q_star)
+
+    print('Observational arc:')
+    print('Number of observations: ', len(obs_arr_ls))
+    print('First observation (UTC) : ', Time(tv_star[0], format='jd').iso)
+    print('Last observation (UTC) : ', Time(tv_star[-1], format='jd').iso)
+
+    n_num = meanmotion(mu_Sun, Q_ls.x[0])
+
+    print('\nORBITAL ELEMENTS (ECLIPTIC, MEAN J2000.0): a, e, taup, omega, I, Omega, T')
+    print('Semi-major axis (a):                 ', Q_ls.x[0], 'au')
+    print('Eccentricity (e):                    ', Q_ls.x[1])
+    print('Time of pericenter passage (tau):    ', Time(Q_ls.x[2], format='jd').iso, 'JDTDB')
+    print('Pericenter distance (q):             ', Q_ls.x[0]*(1.0-Q_ls.x[1]), 'au')
+    print('Apocenter distance (Q):              ', Q_ls.x[0]*(1.0+Q_ls.x[1]), 'au')
+    print('Argument of pericenter (omega):      ', np.rad2deg(Q_ls.x[3]), 'deg')
+    print('Inclination (I):                     ', np.rad2deg(Q_ls.x[4]), 'deg')
+    print('Longitude of Ascending Node (Omega): ', np.rad2deg(Q_ls.x[5]), 'deg')
+    print('Orbital period (T):                  ', 2.0*np.pi/n_num, 'days')
+
+    # PLOT
+    if plot:
+        ra_res_vec = np.rad2deg(rv_star[0::2])*(3600.0)
+        dec_res_vec = np.rad2deg(rv_star[1::2])*(3600.0)
+
+        f, axarr = plt.subplots(2, sharex=True)
+        axarr[0].set_title('Gauss + LS fit residuals: RA, Dec')
+        axarr[0].scatter(tv_star-tv_star[0], ra_res_vec, label='delta RA (\")', marker='+')
+        axarr[0].set_ylabel('RA (\")')
+        axarr[1].scatter(tv_star-tv_star[0], dec_res_vec, label='delta Dec (\")', marker='+')
+        axarr[1].set_xlabel('time (TDB days since first obs)')
+        axarr[1].set_ylabel('Dec (\")')
+        plt.show()
+
+        npoints = 500 # number of points in orbit
+        theta_vec = np.linspace(0.0, 2.0*np.pi, npoints)
+        t_Ea_vec = np.linspace(tv_star[0], tv_star[-1], npoints)
+        x_orb_vec = np.zeros((npoints,))
+        y_orb_vec = np.zeros((npoints,))
+        z_orb_vec = np.zeros((npoints,))
+        x_Ea_orb_vec = np.zeros((npoints,))
+        y_Ea_orb_vec = np.zeros((npoints,))
+        z_Ea_orb_vec = np.zeros((npoints,))
+
+        for i in range(0,npoints):
+            x_orb_vec[i], y_orb_vec[i], z_orb_vec[i] = xyz_frame2(Q_ls.x[0], Q_ls.x[1], theta_vec[i], Q_ls.x[3], Q_ls.x[4], Q_ls.x[5])
+            xyz_Ea_orb_vec_equat = earth_ephemeris(t_Ea_vec[i])/au
+            xyz_Ea_orb_vec_eclip = np.matmul(rot_equat_to_eclip, xyz_Ea_orb_vec_equat)
+            x_Ea_orb_vec[i], y_Ea_orb_vec[i], z_Ea_orb_vec[i] = xyz_Ea_orb_vec_eclip
+
+        ax = plt.axes(aspect='equal', projection='3d')
+
+        # Sun-centered orbits: Computed orbit and Earth's
+        ax.scatter3D(0.0, 0.0, 0.0, color='yellow', label='Sun')
+        ax.plot3D(x_Ea_orb_vec, y_Ea_orb_vec, z_Ea_orb_vec, color='blue', linewidth=0.5, label='Earth orbit')
+        ax.plot3D(x_orb_vec, y_orb_vec, z_orb_vec, 'red', linewidth=0.5, label=body_name_str+' orbit')
+        plt.legend()
+        ax.set_xlabel('x (au)')
+        ax.set_ylabel('y (au)')
+        ax.set_zlabel('z (au)')
+        xy_plot_abs_max = np.max((np.amax(np.abs(ax.get_xlim())), np.amax(np.abs(ax.get_ylim()))))
+        ax.set_xlim(-xy_plot_abs_max, xy_plot_abs_max)
+        ax.set_ylim(-xy_plot_abs_max, xy_plot_abs_max)
+        ax.set_zlim(-xy_plot_abs_max, xy_plot_abs_max)
+        ax.legend(loc='center left', bbox_to_anchor=(1.04,0.5)) #, ncol=3)
+        ax.set_title('Angles-only orbit determ. (Gauss+LS): '+body_name_str)
+        plt.show()
+
+    return Q_ls.x[0], Q_ls.x[1], Time(Q_ls.x[2], format='jd'), np.rad2deg(Q_ls.x[3]), np.rad2deg(Q_ls.x[4]), np.rad2deg(Q_ls.x[5]), 2.0*np.pi/n_num
 
 if __name__ == "__main__":
+    # Earth's mass parameter in appropriate units:
+    mu_Earth = 398600.435436E9 # m^3/seg^2
+    #Earth's radius in appropriate units:
+    # R_Earth =  6378136.3 #m
+    #minimal acceptable altitude for satellites (150 km)??
+    #maximal acceptable altitude for satellites (150 km)??
 
-    data0 = read_file(fname)
+    #write file name of data:
+    fname = '../orbit.csv'
+
+    # load observational data:
+    data = np.loadtxt(fname,skiprows=1,usecols=(0,1,2,3))
+    # generate vector of initial guess of orbital elements:
+    # values written below correspond to solution of ellipse_fit.py for the same file
+    data0 = __read_file(fname)
     kep0, res0 = determine_kep(data0)
 
     a_ = kep0[0][0] # m
